backend/requirements.txt
# Vercel production requirements — only runtime dependencies
# (Dev tools like flake8, mypy, pytest are excluded to keep build fast)

aiohttp==3.13.3
anyio==4.12.0
bcrypt==4.1.3
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.110.1
google-auth==2.47.0
google-auth-httplib2==0.3.0
httpx==0.28.1
motor==3.3.1
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.5.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
requests==2.32.5
requests-oauthlib==2.0.0
starlette==0.37.2
stripe==14.1.0
uvicorn==0.25.0
